% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrupa_nominal.R
\name{agrupa_nominal}
\alias{agrupa_nominal}
\title{agrupa_nominal}
\usage{
agrupa_nominal(tbla, variable_name, target_name, limite, symbol_to_split)
}
\arguments{
\item{tbla}{table with data. It has to have the nominal variable and the target variable.}

\item{variable_name}{name of the ordinal variable that you want to group.}

\item{target_name}{name if the target variable.}

\item{limite}{probability of not belonging to the same group. Used for the hypergeometric test.}

\item{symbol_to_split}{symbol to split values of variables}
}
\description{
group values of nominal varibles according to a target variable
.
}
\examples{
set.seed(1)
tbla<-data.frame(grupo=rep(c('a','b','c','d','e'),100/5),valor=as.numeric(runif(100)>.5 ))
agrupa_nominal(tbla, variable_name='grupo', target_name='valor', limite=0.05)
}
