% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrupa_nominal_filtra_small.R
\name{agrupa_nominal_filtra_small}
\alias{agrupa_nominal_filtra_small}
\title{agrupa_nominal_filtra_small}
\usage{
agrupa_nominal_filtra_small(tbla, target_name, variable_name, limite,
  symbol_to_split = "\%-\%", limite_grupo = 100)
}
\arguments{
\item{tbla}{table with data. It has to have the nominal variable and the target variable.}

\item{target_name}{name if the target variable.}

\item{variable_name}{name of the ordinal variable that you want to group.}

\item{limite}{probability of not belonging to the same group. Used for the hypergeometric test.}

\item{symbol_to_split}{symbol to split values of variables}

\item{limite_grupo}{min number for each group to be analized}
}
\description{
group values of nominal variables according to a target variable
.
}
\examples{
set.seed(1)
tbla<-rbind(data.frame(grupo=rep(c('a','b','c','d','e'),100/5),valor=as.numeric(runif(100)>.5 )),data.frame(grupo=rep(c('c','d'),100/2),valor=as.numeric(runif(100)>.5 )))
agrupa_nominal_filtra_small(tbla, target_name='valor', variable_name='grupo',limite=0.05, symbol_to_split='\%#\%', limite_grupo=50)
}
