% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrupa_ordinal.R
\name{agrupa_ordinal}
\alias{agrupa_ordinal}
\title{agrupa_ordinal}
\usage{
agrupa_ordinal(tbla_0, target_name, variable_name, grupos_iniciales,
  limite)
}
\arguments{
\item{variable_name}{name of the ordinal variable that you want to group.}

\item{grupos_iniciales}{number of quantiles that you want to create. Uses quantile function.}

\item{limite}{probability of not belonging to the same group. Used for the hypergeometric test.}

\item{tbla}{table with data. It has to have the ordinal variable and the target variable.}

\item{target-name}{name if the target variable.}
}
\description{
group values of ordinal varibles according to a target variable
.
}
\examples{
set.seed(1)
valor_variable1=round(runif(100)*10,0)
target_variable1=as.numeric((valor_variable1/10+runif(100)*0.25) >0.525)
tbla<-data.frame(variable1=valor_variable1, target_n= target_variable1)
agrupa_ordinal(tbla,  'target_n','variable1', 10, 0.8)
}
