% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplica_agrupa_nominal.R
\name{aplica_agrupa_nominal}
\alias{aplica_agrupa_nominal}
\title{aplica_agrupa_nominal}
\usage{
aplica_agrupa_nominal(train_tbl, char_cols, target_name, lim_cant_categ,
  limite, symbol_to_split)
}
\arguments{
\item{train_tbl}{table with data. It has to have the nominal variable and the target variable.}

\item{char_cols}{name of the nominal variables that you want to group.}

\item{target_name}{name if the target variable.}

\item{lim_categ}{probability of not belonging to the same group. Used for the hypergeometric test.}
}
\description{
apply agrupa_nominal to a list of variables
}
\examples{
tbla<-data.frame(
grupo1=rep(c('a','b','c','d','e'),100/5),
valor=c( as.numeric(rep(runif(90), 1)>.5) , rep(0, 5), rep(1, 5)) ,
grupo2=rep(c('aa','bb','cc','cc','cc'),100/5)
)
 aplica_agrupa_nominal(train_tbl=tbla, char_cols=c('grupo1', 'grupo2'), target_name='valor', lim_cant_categ=20, limite=0.05, symbol_to_split = '\%#\%')
}
