% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplica_agrupa_ordinal.R
\name{aplica_agrupa_ordinal}
\alias{aplica_agrupa_ordinal}
\title{aplica_agrupa_ordinal}
\usage{
aplica_agrupa_ordinal(train_tbl, num_cols, target_name, num_ini_groups,
  limite)
}
\arguments{
\item{train_tbl}{table with data. It has to have the nominal variable and the target variable.}

\item{num_cols}{name of the numeric variables that you want to group.}

\item{target_name}{name of the target variable.}

\item{num_ini_groups}{probability of not belonging to the same group. Used for the hypergeometric test.}

\item{lim_categ}{probability of not belonging to the same group. Used for the hypergeometric test.}
}
\description{
apply agrupa_ordinal to a list of variables
}
\examples{
tbla<-data.frame(
grupo1=rep(1:5,100/5),
valor=c( as.numeric(rep(runif(90), 1)>.5) , rep(0, 5), rep(1, 5)) ,
grupo2=rep(c(11,22,33,33,33),100/5))
aplica_agrupa_ordinal(train_tbl=tbla, num_cols=c('grupo1', 'grupo2'), target_name='valor', num_ini_groups=20, limite=0.05)
}
