% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplica_cortes_ordinales.R
\name{aplica_cortes_ordinales}
\alias{aplica_cortes_ordinales}
\title{aplica_cortes_ordinales}
\usage{
aplica_cortes_ordinales(tbla_cortes_variables_ord, tbl, campo_merge)
}
\arguments{
\item{tbla_cortes_variables_ord}{table with cuts return by aplica_agrupa_nominal.}

\item{tbl}{table with data to apply the cuts.}

\item{campo_merge}{name of the variable used to merge the result and the data (index).}
}
\description{
apply the cuts obteined with aplica_agrupa_nominal to a list of new variables
.
}
