% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devuelve_tabla_agrupada.R
\name{devuelve_tabla_agrupada}
\alias{devuelve_tabla_agrupada}
\title{devuelve_tabla_agrupada}
\usage{
devuelve_tabla_agrupada(tbla, variable_nom, target_nom)
}
\arguments{
\item{tbla}{table with data. It has to have the nominal variable and the target variable.}

\item{variable_nom}{name of the ordinal variable that you want to group.}

\item{target_nom}{name if the target variable.}
}
\description{
group values of nominal variables
}
\examples{
set.seed(1)
}
