% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estabilidad_poder_predictivo_variable.R
\name{estabilidad_poder_predictivo_de_variable}
\alias{estabilidad_poder_predictivo_de_variable}
\title{performance_bivariado_rpart}
\usage{
estabilidad_poder_predictivo_de_variable(tbla, variable_name, target_name,
  filtros_train, var_particion_estabilidad)
}
\arguments{
\item{tbla}{table with data. It has to have the variable and the target variable.}

\item{variable_name}{name of the variable that you want to analyze.}

\item{target-name}{name of the target variable.}
}
\description{
group values of ordinal varibles according to a target variable
.
}
\examples{
set.seed(1)
x1 = rnorm(1000)
x2 = rnorm(1000)
x3= ifelse(as.factor(x2>0.5)==T, 'A', 'B')
z = 1 + 2 * x1 + 3 * x2
pr = 1/(1+exp(-z))
y = rbinom(1000,1,pr)
tbla = data.frame(y=y,x1=x1,x2=x2, x3=x3)
filtros_train= (tbla$random=runif(nrow(tbla)))<0.5
estabilidad_poder_predictivo_de_variable (tbla, variable_name='x3',target_name='y', filtros_train= )
estabilidad_poder_predictivo_de_variable (tbla, variable_name='x2',target_name='y',filtros_train )
}
