% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agrupa_ordinal_internal_funcs.R
\name{genera_deciles}
\alias{genera_deciles}
\title{#' @import dplyr}
\usage{
genera_deciles(tbla, variable_name, target_name, q)
}
\description{
devtools::use_package('dplyr')
`%>%`<- dplyr::`%>%`
}
