% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_performance_logistica.R
\name{loop_performance_logistica}
\alias{loop_performance_logistica}
\title{loop_performance_logistica}
\usage{
loop_performance_logistica(tbla, lista, target_name, flag_numerica = 0,
  limite_categ = 100, limite_steps = 500)
}
\arguments{
\item{tbla}{table with data. It has to have the variables and the target variable.}

\item{lista}{list of names of the variables that you want to analyze.}

\item{limite_categ}{max number of categories of the variable to be considered.}

\item{limite_steps}{max number of steps for the stepAIC.}

\item{target-name}{name of the target variable.}
}
\description{
bivariate analysis
.
}
\examples{
set.seed(1)
x1 = rnorm(1000)
x2 = rnorm(1000)
z = 1 + 2 \\* x1 + 3 \\* x2 ^2
pr = 1/(1+exp(-z))
y = rbinom(1000,1,pr)
tbla = data.frame(y=y,x1=x1,x2=x2)
loop_performance_logistica (tbla, lista=c('x1', 'x2'),target_name='y', flag_numerica=1 )
loop_performance_logistica (tbla, lista=c('x3'),target_name='y' , flag_numerica=0)
}
