% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loop_performance_rpart.R
\name{loop_performance_rpart}
\alias{loop_performance_rpart}
\title{loop_performance_rpart}
\usage{
loop_performance_rpart(tbla, lista, target_name)
}
\arguments{
\item{tbla}{table with data. It has to have the variables and the target variable.}

\item{lista}{list of names of the variables that you want to analyze.}

\item{target-name}{name of the target variable.}
}
\description{
bivariate analysis
.
}
\examples{
set.seed(1)
x1 = rnorm(1000)
x2 = rnorm(1000)
z = 1 + 2 \\* x1 + 3 \\* x2
pr = 1/(1+exp(-z))
y = rbinom(1000,1,pr)
tbla = data.frame(y=y,x1=x1,x2=x2)
loop_performance_rpart (tbla, lista=c('x1', 'x2'),target_name='y' )
}
