% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/non_overlapping_sample.R
\name{non_overlapping_sample}
\alias{non_overlapping_sample}
\title{non_overlapping_sample
take a sample with non overlapping elements for the same id (i.e. client id) according to a window of days.}
\usage{
non_overlapping_sample(tbla, date_field, id_field, window_days)
}
\arguments{
\item{tbla}{table with data. It has to have the date field and the id field (could be client id).}

\item{date_field}{name of the  date field used to create the window of days.}

\item{id_field}{name of the id (client id).}

\item{window_days}{number of days that are going to be in each side of the selected date in the sample.}
}
\description{
non_overlapping_sample
take a sample with non overlapping elements for the same id (i.e. client id) according to a window of days.
}
\examples{
set.seed(1)
sem = sample(seq.Date(ymd(20150101),ymd(20180101),1), 3000, replace = T)
base = data.frame(fc_fin_semana = sem, cd_cliente=round(runif(3000)*10,0))
base=base[!duplicated(base),]
non_overlapping_sample(base, date_field='fc_fin_semana', 'cd_cliente', 90)
}
