% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance_modelo_logistica.R
\name{performance_modelo_logistica}
\alias{performance_modelo_logistica}
\title{performance_bivariado_transf_log}
\usage{
performance_modelo_logistica(tbla2, mod_all, variable_name, form_all,
  limite_steps)
}
\arguments{
\item{mod_all}{formula}

\item{tbla}{table with data. It has to have the variable and the target variable.}
}
\description{
group values of ordinal varibles according to a target variable
.
}
