% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quita_char_especiales.R
\name{quita_char_especiales}
\alias{quita_char_especiales}
\title{quita_char_especiales}
\usage{
quita_char_especiales(tbla, campo)
}
\arguments{
\item{tbla}{table with data. It has to have the variable that you want to clean.}

\item{campo}{name of the variable that you want to clean.}
}
\description{
remove special characters from text
.
}
\examples{

}
