% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine_level_0.R
\name{redefine_level_0}
\alias{redefine_level_0}
\title{redefine_level_0}
\usage{
redefine_level_0(df_agrupada_y, variables, nombre_target)
}
\arguments{
\item{df_agrupada_y}{table with data. It has to have the nominal variable and the target variable.}

\item{variables}{names of the factor variables.}

\item{nombre_target}{name of the target variable.}
}
\description{
redefine level 0 for a list of factors variables
.
}
\examples{
x1 = rnorm(1000)
x2 = rnorm(1000)
x3= ifelse(as.factor(x2>0.5)==T, 'A', 'B')
x4= ifelse(as.factor(x2>0.7)==T, 'C', 'D')
z = 1 + 2 * x1 + 3 * x2+2*x4
pr = 1/(1+exp(-z))
y = rbinom(1000,1,pr)
tbla = data.frame(y=y,x1=x1,x2=x2, x3=x3, x4=x4)
tbla<-redefine_level_0( df_agrupada_y=tbla ,variables=c('x3',  'x4') ,nombre_target='y')
}
