% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_cruzada_br.R
\name{tabla_cruzada_br}
\alias{tabla_cruzada_br}
\title{tabla_cruzada_br.}
\usage{
tabla_cruzada_br(tbla, target_name, variable_name1, variable_name2)
}
\arguments{
\item{tbla}{table with all data}

\item{target_name}{table with data}

\item{variable_name1}{table with data}

\item{variable_name2}{table with data}
}
\description{
tabla_cruzada_br.
}
