% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabla_estimadores.R
\name{tabla_estimadores}
\alias{tabla_estimadores}
\title{tabla_estimadores}
\usage{
tabla_estimadores(lr)
}
\arguments{
\item{lr}{model}
}
\description{
creates table of the summary of a model with all the levels and the importance of the variables
.
}
\examples{
x1 = rnorm(1000)
x2 = rnorm(1000)
x3= ifelse(as.factor(x2>0.5)==T, 'A', 'B')
x4= ifelse(as.factor(x2>0.7)==T, 'C', 'D')
z = 1 + 2 * x1 + 3 * x2+2*x4
pr = 1/(1+exp(-z))
y = rbinom(1000,1,pr)
tbla = data.frame(y=y,x1=x1,x2=x2, x3=x3, x4=x4)
tbla<-redefine_level_0( df_agrupada_y=tbla ,variables=c('x3',  'x4') ,nombre_target='y')
filtros_train= (tbla$random=runif(nrow(tbla)))<0.5
lr <- glm(f, tbla[ filtros_train, ], family = 'binomial')
tabla_estimadores(lr)
}
