% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aplica_agrupa_ordinal.R
\name{aplica_agrupa_ordinal}
\alias{aplica_agrupa_ordinal}
\title{aplica_agrupa_ordinal}
\usage{
aplica_agrupa_ordinal(train_tbl, num_cols, target_name, num_ini_groups,
  limite)
}
\arguments{
\item{train_tbl}{table with data. It has to have the nominal variable and the target variable.}

\item{num_cols}{name of the numeric variables that you want to group.}

\item{target_name}{name if the target variable.}

\item{num_ini_groups}{probability of not belonging to the same group. Used for the hypergeometric test.}

\item{lim_categ}{probability of not belonging to the same group. Used for the hypergeometric test.}
}
\description{
apply agrupa_ordinal to a list of variables
.
}
