% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/redefine_level_0.R
\name{redefine_level_0}
\alias{redefine_level_0}
\title{redefine_level_0}
\usage{
redefine_level_0(df_agrupada_y, variables, nombre_target)
}
\arguments{
\item{df_agrupada_y}{table with data. It has to have the nominal variable and the target variable.}

\item{variables}{names of the factor variables.}

\item{nombre_target}{name of the target variable.}
}
\description{
redefine level 0 for a list of factors variables
.
}
